[polybool2] -Fix: prints wrapper: when printing vector coords as %ld, cast them to long, rnd_coord_t can be anything
git-svn-id: svn://repo.hu/librnd/trunk@37397 1d61de99-560f-4dba-8957-1e505d42358b

+++ src/librnd/polybool2/pa_prints.h
@@ -27,7 +27,7 @@ RND_INLINE long P_print_vect_(rnd_vector_t v, Pfmt_t *format, void (*put_ch)(voi
 	if (pa_coord_is_mm(v[0]) || pa_coord_is_mm(v[1]))
 		len = rnd_sprintf(tmp, "%.012mm;%.012mm", v[0], v[1]);
 	else
-		len = rnd_sprintf(tmp, "%ld;%ld", v[0], v[1]);
+		len = rnd_sprintf(tmp, "%ld;%ld", (long)v[0], (long)v[1]);
 
 	P_PRINT_STR(tmp);
